+++ test/keypin/core_test.clj
@@ -40,7 +40,23 @@
   (testing "duration?"
     (is (ku/duration? (kt/->Duration 10 TimeUnit/MILLISECONDS)))
     (is (ku/duration? [10 TimeUnit/MILLISECONDS]))
-    (is (ku/duration? {:time 10 :unit TimeUnit/MILLISECONDS}))))
+    (is (ku/duration? [10 :millis]))
+    (is (ku/duration? {:time 10 :unit TimeUnit/MILLISECONDS}))
+    (is (ku/duration? {:time 10 :unit :millis}))
+    (is (= 10 (kt/dur-time (kt/->Duration 10 TimeUnit/MILLISECONDS))))
+    (is (= 10 (kt/dur-time [10 :millis])))
+    (is (= 10 (kt/dur-time {:time 10 :unit :millis})))
+    (is (= TimeUnit/MILLISECONDS (kt/dur-unit (kt/->Duration 10 TimeUnit/MILLISECONDS))))
+    (is (= TimeUnit/MILLISECONDS (kt/dur-unit [10 :millis])))
+    (is (= TimeUnit/MILLISECONDS (kt/dur-unit {:time 10 :unit :millis})))
+    (is (= (kt/millis (kt/->Duration 10 TimeUnit/MILLISECONDS))
+          (kt/millis [10 TimeUnit/MILLISECONDS])
+          (kt/millis [10 :millis])
+          (kt/millis {:time 10 :unit TimeUnit/MILLISECONDS})
+          (kt/millis {:time 10 :unit :millis}))))
+  (testing "atom?"
+    (is (ku/atom? (atom nil)))
+    (is (ku/atom? (atom :re)))))
 
 
 (deftest test-config-file-reader
@@ -285,6 +301,8 @@
   (is (= [34 TimeUnit/MILLISECONDS] (ku/any->duration "foo" "34ms")))
   (is (ku/duration? (ku/any->duration "foo" [34 :ms])))
   (is (= [34 TimeUnit/MILLISECONDS] (ku/any->duration "foo" [34 :ms])))
+  (is (= [34 TimeUnit/MILLISECONDS] (ku/any->duration "foo" [34 :milliseconds])))
+  (is (= [34 TimeUnit/MILLISECONDS] (ku/any->duration "foo" [34 :millis])))
   ;; collection parsers
   (is (= ["foo" "bar" "baz"]
         (ku/any->vec "foo" ["foo" "bar" "baz"])))